requirements_new.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
billiard==4.2.2
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.2
cycler==0.12.1
Django==4.2.24
django-cors-headers==4.4.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.183.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
httplib2==0.31.0
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.9
kombu==5.5.4
MarkupSafe==3.0.3
matplotlib==3.10.7
networkx==3.5
numpy==2.3.4
oauthlib==3.3.1
packaging==25.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==5.29.5
psycopg==3.2.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
PyJWT==2.9.0
pyparsing==3.1.4
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.3
redis==6.4.0
regex==2025.9.18
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.14
